Fix error test
